improve styles + semantic tags + rename component
