+++ package.json
@@ -18,6 +18,7 @@
     "@reduxjs/toolkit": "^1.9.5",
     "axios": "^1.4.0",
     "bootstrap": "^5.3.0",
+    "classnames": "^2.3.2",
     "compose-function": "^3.0.3",
     "framer-motion": "^10.15.0",
     "i18next": "^23.4.1",
