+++ src/app/styles/index.scss
@@ -60,35 +60,34 @@
    }
 }
 
-
 .catalog-icon:hover circle {
-    fill: #00BCD4;
- }
+    @apply fill-accent-turkus;
+}
 
- .present-icon:hover path:nth-child(1){
-    fill: #4495ff;
- }
+.present-icon {
+&:hover {
+    path:nth-child(1){
+        @apply fill-[#4495ff];
+    }
 
- .present-icon:hover path:nth-child(2), .present-icon:hover path:nth-child(5){
-    fill: #006eff;
- }
+    path:nth-child(3) {
+        @apply fill-accent-turkus;
+    }        
+    
+    path:nth-child(2), path:nth-child(5){
+        @apply fill-[#006eff];
+    }
+}
+}
+ 
+.search-icon:focus-within path {
+   @apply fill-[#0047FF];
+}
 
- .present-icon:hover path:nth-child(3) {
-    fill: #00BCD4;
- }
-
- .search-icon:focus-within path {
-   fill: #0047FF;
- }
-
-.user-icon:hover path {
-   fill: white;
+.user-icon:hover path, .basket-icon:hover path {
+   @apply fill-white;
 }
 
 .wishlist-icon:hover path {
-   fill: #FF3232;
-}
-
-.basket-icon:hover path {
-   fill: white;
+   @apply fill-accent-red;
 }
