+++ src/shared/assets/svg/Basket.tsx
