+++ src/shared/assets/svg/BlueClose.tsx
