+++ src/shared/assets/svg/Catalog.tsx
