+++ src/shared/assets/svg/Present.tsx
