+++ src/shared/assets/svg/Search.tsx
