+++ src/shared/assets/svg/UserAccount.tsx
