+++ src/shared/assets/svg/Wishlist.tsx
