+++ src/shared/index.ts
@@ -1,4 +1,12 @@
 /** Assets */
+export { Basket } from "./assets/svg/Basket";
+export { BlueClose } from "./assets/svg/BlueClose";
+export { Catalog } from "./assets/svg/Catalog";
+export { DropDown } from "./assets/svg/DropDown";
+export { Present } from "./assets/svg/Present";
+export { Search } from "./assets/svg/Search";
+export { UserAccount } from "./assets/svg/UserAccount";
+export { Wishlist } from "./assets/svg/Wishlist";
 
 /** Hooks */
 export { useNetwork } from "./hooks/useNetwork";
@@ -15,13 +23,3 @@ export { ADMIN, MANAGER, GUEST_USER, AUTH_USER } from "./constants/roles";
 /** UI  */
 export { default as Preloader } from "./UI/Preloader";
 export { default as Logo } from "./UI/Logo";
-
-/** Icons  */
-export { Basket } from "./icons/Basket";
-export { BlueClose } from "./icons/BlueClose";
-export { Catalog } from "./icons/Catalog";
-export { DropDown } from "./icons/DropDown";
-export { Present } from "./icons/Present";
-export { Search } from "./icons/Search";
-export { UserAccount } from "./icons/UserAccount";
-export { Wishlist } from "./icons/Wishlist";
